Removing duplicated empty state code and unifying it in bookmarks search and history list
