+++ Client/Frontend/Browser/Toolbars/BottomToolbar/Menu/Bookmarks/BookmarksViewController.swift
@@ -102,7 +102,7 @@ class BookmarksViewController: SiteTableViewController, ToolbarUrlActionsProtoco
     private var isBookmarksBeingSearched = false
     private let bookmarksSearchController = UISearchController(searchResultsController: nil)
     private var bookmarksSearchQuery = ""
-    private lazy var noSearchResultOverlayView = createNoSearchResultOverlayView()
+    private lazy var noSearchResultOverlayView = EmptyStateOverlayView(description: Strings.noSearchResultsfound)
 
     // MARK: Lifecycle
     
@@ -232,35 +232,7 @@ class BookmarksViewController: SiteTableViewController, ToolbarUrlActionsProtoco
     private func updateLastVisitedFolder(_ folder: Bookmarkv2?) {
         Preferences.Chromium.lastBookmarksFolderNodeId.value = folder?.objectID ?? -1
     }
-    
-    private func createNoSearchResultOverlayView() -> UIView {
-        let overlayView = UIView().then {
-            $0.backgroundColor = .secondaryBraveBackground
-        }
-        
-        let welcomeLabel = UILabel().then {
-            $0.text = Strings.noSearchResultsfound
-            $0.textAlignment = .center
-            $0.font = DynamicFontHelper.defaultHelper.DeviceFontLight
-            $0.textColor = .braveLabel
-            $0.numberOfLines = 0
-            $0.adjustsFontSizeToFitWidth = true
-        }
-        
-        overlayView.addSubview(welcomeLabel)
-        
-        welcomeLabel.snp.makeConstraints { make in
-            make.centerX.equalTo(overlayView)
-            // Sets proper top constraint for iPhone 6 in portait and for iPad.
-            make.centerY.equalTo(overlayView).offset(-180).priority(100)
-            // Sets proper top constraint for iPhone 4, 5 in portrait.
-            make.top.greaterThanOrEqualTo(overlayView).offset(50)
-            make.width.equalTo(170)
-        }
-        
-        return overlayView
-    }
-    
+
     private func updateEmptyPanelState() {
         if isBookmarksBeingSearched, bookmarkManager.fetchedSearchObjectsCount == 0 {
             showEmptyPanelState()
