+++ Client/Frontend/Browser/Toolbars/BottomToolbar/Menu/HistoryViewController.swift
@@ -14,7 +14,11 @@ class HistoryViewController: SiteTableViewController, ToolbarUrlActionsProtocol 
     
     weak var toolbarUrlActionsDelegate: ToolbarUrlActionsDelegate?
 
-    private lazy var emptyStateOverlayView = HistoryStateOverlayView()
+    private lazy var emptyStateOverlayView = EmptyStateOverlayView(
+        description: Preferences.Privacy.privateBrowsingOnly.value
+            ? Strings.History.historyPrivateModeOnlyStateTitle
+            : Strings.History.historyEmptyStateTitle,
+        icon: #imageLiteral(resourceName: "emptyHistory"))
 
     private let spinner = UIActivityIndicatorView().then {
         $0.snp.makeConstraints { make in
@@ -415,57 +419,5 @@ extension HistoryViewController: UISearchControllerDelegate {
         
         isHistoryBeingSearched = false
         tableView.reloadData()
-    }
-}
-
-class HistoryStateOverlayView: UIView {
-    
-    private let logoImageView = UIImageView(image: #imageLiteral(resourceName: "emptyHistory").template).then {
-        $0.tintColor = .braveLabel
-    }
-    
-    private let informationLabel = UILabel().then {
-        $0.text = Preferences.Privacy.privateBrowsingOnly.value
-            ? Strings.History.historyPrivateModeOnlyStateTitle
-            : Strings.History.historyEmptyStateTitle
-        $0.textAlignment = .center
-        $0.font = DynamicFontHelper.defaultHelper.DeviceFontLight
-        $0.textColor = .braveLabel
-        $0.numberOfLines = 0
-        $0.adjustsFontSizeToFitWidth = true
-    }
-    
-    required init() {
-        super.init(frame: .zero)
-        
-        backgroundColor = .secondaryBraveBackground
-
-        addSubview(logoImageView)
-        
-        logoImageView.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.size.equalTo(60)
-            // Sets proper top constraint for iPhone 6 in portait and for iPad.
-            make.centerY.equalToSuperview().offset(-180).priority(100)
-            // Sets proper top constraint for iPhone 4, 5 in portrait.
-            make.top.greaterThanOrEqualToSuperview().offset(50)
-        }
-        
-        addSubview(informationLabel)
-        
-        informationLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(logoImageView.snp.bottom).offset(15)
-            make.width.equalTo(170)
-        }
-    }
-    
-    @available(*, unavailable)
-    required init(coder: NSCoder) {
-        fatalError()
-    }
-    
-    func updateInfoLabel(with text: String) {
-        informationLabel.text = text
     }
 }
